AFDP-5305 just use 'long' all the way... don't long-to-int-to-long
Former-commit-id: f339fbe69846a16914737a37a17bcab272b22149

+++ acm-services/acm-service-ecm/src/main/java/com/armedia/acm/plugins/ecm/pipeline/presave/EcmFileNewMetadataHandler.java
@@ -56,10 +56,10 @@ public class EcmFileNewMetadataHandler implements PipelineHandler<EcmFile, EcmFi
             version.setVersionTag(cmisDocument.getVersionLabel());
             version.setVersionMimeType(entity.getFileActiveVersionMimeType());
             version.setVersionFileNameExtension(entity.getFileActiveVersionNameExtension());
-            int fileSizeBytes = pipelineContext.getMergedFileByteArray() != null &&
+            long fileSizeBytes = pipelineContext.getMergedFileByteArray() != null &&
                     pipelineContext.getMergedFileByteArray().length > 0 ? pipelineContext.getMergedFileByteArray().length
-                            : pipelineContext.getFileContents() != null ? (int) pipelineContext.getFileContents().length() : 0;
-            version.setFileSizeBytes(Long.valueOf(fileSizeBytes));
+                            : pipelineContext.getFileContents() != null ? pipelineContext.getFileContents().length() : 0;
+            version.setFileSizeBytes(fileSizeBytes);
 
             // file metadata
             if (pipelineContext.getDetectedFileMetadata() != null)
